refactor: use new module hierarchy for imports
