+++ hcloud/actions/client.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import time
 from typing import TYPE_CHECKING, NamedTuple
 
-from ..core.client import BoundModelBase, ClientEntityBase
-from ..core.domain import Meta
+from ..core import BoundModelBase, ClientEntityBase
+from ..core import Meta
 from .domain import Action, ActionFailedException, ActionTimeoutException
 
 if TYPE_CHECKING:
