+++ hcloud/floating_ips/client.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, NamedTuple
 
-from ..actions.client import ActionsPageResult, BoundAction
-from ..core.client import BoundModelBase, ClientEntityBase, GetEntityByNameMixin
-from ..core.domain import Meta
-from ..locations.client import BoundLocation
+from ..actions import ActionsPageResult, BoundAction
+from ..core import BoundModelBase, ClientEntityBase, GetEntityByNameMixin
+from ..core import Meta
+from ..locations import BoundLocation
 from .domain import CreateFloatingIPResponse, FloatingIP
 
 if TYPE_CHECKING:
@@ -18,7 +18,7 @@ class BoundFloatingIP(BoundModelBase):
     model = FloatingIP
 
     def __init__(self, client, data, complete=True):
-        from ..servers.client import BoundServer
+        from ..servers import BoundServer
 
         server = data.get("server")
         if server is not None:
